backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
gunicorn>=21.2.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
motor>=3.3.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-multipart>=0.0.9
google-cloud-storage>=2.14.0
google-genai>=0.3.0
pikepdf>=8.0.0
pypdf>=4.0.0
pymupdf>=1.23.8
python-magic-bin>=0.4.14 ; platform_system == "Windows"
python-magic>=0.4.27 ; platform_system != "Windows"
reportlab>=4.0.9
python-docx>=1.1.0
email-validator>=2.1.0.post1
celery>=5.3.6
redis>=5.0.1
eventlet>=0.33.3
prometheus-client>=0.19.0